Fix for seaonalstats

+++ src/main/kotlin/de/jan/r6statsjava/R6GameTime.kt
@@ -2,7 +2,7 @@ package de.jan.r6statsjava
 
 class R6GameTime(val seconds: Int) {
 
-    fun toMilliseconds() : Long {
+    fun toMilliseconds() : Int {
         return seconds * 1000
     }
